Move compat code to seccomp_compat.{c,h}
This separates code sensitive to libseccomp C headers version.

While at it, fix the comment telling about the minimally required
libseccomp version.

No change in functionality.
